requirements.txt
# CLIP embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector store
chromadb>=0.4.0

# Face recognition
face_recognition>=1.3.0
dlib>=19.24.0

# Image processing
Pillow>=10.0.0

# HTTP (for Ollama)
requests>=2.31.0

# Utilities
numpy>=1.24.0
